+++ src/com/intellij/idea/plugin/hybris/impex/lang/folding/ImpexFoldingBuilder.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,6 +17,7 @@
  */
 package com.intellij.idea.plugin.hybris.impex.lang.folding
 
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexFile
 import com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils
 import com.intellij.lang.ASTNode
 import com.intellij.lang.folding.FoldingDescriptor
@@ -58,11 +59,9 @@ class ImpexFoldingBuilder : AbstractImpExFoldingBuilder() {
         val text = ImpexFoldingPlaceholderBuilderFactory.getPlaceholderBuilder(psi.project).getPlaceholder(psi)
         var resolvedMacro = text
         if (text.startsWith("$")) {
-            val cache = ImpexMacroUtils.getFileCache(psi.containingFile).value
-            val descriptor = cache[text]
-            if (descriptor != null) {
-                resolvedMacro = descriptor.resolvedValue
-            }
+            (psi.containingFile as ImpexFile)
+                .getSuitableMacroDescriptor(text, psi)
+                ?.let { resolvedMacro = it.resolvedValue }
         }
         return if (resolvedMacro.length <= text.length) {
             resolvedMacro